added moved update logic to separate module
