+++ plugins/conversation-insights/src/summary-prompt.ts
@@ -17,7 +17,11 @@ const formatMessages = (
   return {
     role: 'user',
     content: JSON.stringify(
-      { previousTitle: context.previousTitle, previousSummary: context.previousSummary, messages },
+      {
+        previousTitle: context.previousTitle,
+        previousSummary: context.previousSummary,
+        messages: messages.reverse(),
+      },
       null,
       2
     ),
@@ -56,7 +60,7 @@ Instructions:
 - Make the title short and descriptive (few words).
 - Keep the summary concise (one or two sentences).
 - Do not include extra commentary, formatting, or explanation outside the JSON output.
-- Keep in mind that the messages at the beginning of the list are the most recent. The summary should reflect the recent messages more than the previous ones
+- The messages are in order, to the most recent ones are at the end of the list
 
 Example:
